First minor version (#1)
Created some entities of this game:
• Game;
•  Player;
•  Inventory and Item;
•  World;
•  WorldHelper;
•  Storage;
•  Chunk;
•  Block;
•  ChunkMesh;
•  BlockMesh
and other.

The game logic a bit complex. Firstly, in specific positions of crate places singletons: aspect ratio, storage, block renderer and other. With it game logic have less complexity from initial stage by this.

– Blocks, Chunks
Storing all block in hash-table will be costly, and storing in vector exceeds all memory. Therefore exists chunks, which stores blocks by 16x16x216 blocks (for my situation) as array of unsigned longs (u64). My choice for this type based on possibility storing data as bits and u64 can be used as "array of bits". Last 16 bits uses as block data, the other bits are reserved for metadata.

– Rendering
Rendering all block in one chunk always will be expensive, that I don't like. At this point I solved it by specific algorithm, that makes chunk mesh, which contains from block mesh. As you see in code, I used WorldHelper (let's name it WH) for defining, is it block in other neighbor chunk. If the chunk doesn't exists, the code in WH uses specific method of Chunk, that generates only one block and it always will be valid.
And there is texture atlas as .png file. I store it in 2d texture OpenGL array. Every tile in texture atlas represent block or part of block (like wood).

– Unsafety
I know, unsafe code is not good for Rust, because behavior will be undefined. Unfortunately, OpenGL written on C (or C++, I forgot about it) and this language is not safe for Rust. Always.
What about my simple Rust macro lock!, that contains unsafe code for multithreading? Very sad to say it, I not found the code or structure, which is safe to write data from some threads. I know, this code is enough safe for it. If I find a problem with it, I'll solve it.

– Calculus and logic
Based on my simple experience, computation of position or collision is not very simple tasks. In this code you can find some commentaries that explaines, why I wrote this code. Placing or destroying block based on intersection of ray (player-camera taget vector). Placing block have more complex code that destroying, because need define a face of block and shift to need coordinate.
I not add the jump code logic, because it can be complex. It will be in the next minor version.
Currently the player have inventory, but it is useless before adding some logic for picking block. For testing it, you can use method straight in world initialization.

– General information
Some logic and functionalities I omit, because the upper sections are main for focus. The other details is not necessary.